multiples cambios en la db

+++ laravel/app/Http/Controllers/AtazaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Jobs\SyncAtazaToOdoo;
 use App\Models\Ataza;
+use App\Models\User;
 use App\Models\Pisua;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
@@ -36,8 +37,10 @@ class AtazaController extends Controller
      */
     public function create(Pisua $pisua)
     {
+        $usuarios = $pisua->users;
         return Inertia::render('Tasks/CreateTask', [
-            'pisua' => $pisua
+            'pisua' => $pisua,
+            'usuarios' => $usuarios
         ]);
     }
 
@@ -46,18 +49,25 @@ class AtazaController extends Controller
      */
     public function store(Request $request, Pisua $pisua)
     {
-        // 1. Validamos que los datos vengan bien
-        $request->validate([
+        //arduradunak array
+        $validated = $request->validate([
             'izena' => 'required|string|max:255',
-            'egilea' => 'required|string|max:255',
-            'arduraduna' => 'required|string|max:255',
+            'arduradunak' => 'required|array|min:1',
+            'arduradunak.*' => 'exists:users,id',
+            'data' => 'required|date',
+        ]);
+        // ataza sortzen du
+        $ataza = Ataza::create([
+            'izena' => $validated['izena'],
+            'user_id' => $request->arduraduna_id, // para asignar al responsable de task
+            'pisua_id' => $pisua->id,
+            'data' => $validated['data'],
+            'egoera' => 'egiteko',
         ]);
 
-        // 2. Creamos la tarea usando asignación masiva
-        Ataza::create($request->all());
+        $ataza->arduradunak()->attach($validated['arduradunak']);
 
-        // 3. Redireccionamos al listado
-        return redirect()->route('atazak.index')
+        return redirect()->route('atazak.index', $pisua->id)
             ->with('success', 'Ataza ondo gorde da!');
     }
 
@@ -85,9 +95,15 @@ class AtazaController extends Controller
     /**
      * Elimina la tarea.
      */
-    public function destroy(Ataza $ataza)
+public function destroy(Pisua $pisua, Ataza $ataza)
     {
+        //Verificar que la tarea pertenece a ese piso
+        if ($ataza->pisua_id !== $pisua->id) {
+            abort(404);
+        }
+
         $ataza->delete();
+        
         return redirect()->back()->with('success', 'Ataza ezabatu da!');
     }
